+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.1.1 under development
 
 - Enh #150: Cleanup templates, remove legacy code (@vjik)
+- New #151: Add `$traceLink` parameter to `HtmlRenderer` to allow linking to trace files (@vjik)
 
 ## 4.1.0 April 18, 2025